+++ lib/twenv/commands/archive_timeline.rb
@@ -23,7 +23,7 @@ class TWEnv::ArchiveTimeline < TWEnv::Command
                              archive.map(&:id)
     line.end
     complete_archive(path, archive, local_name: "archived_timeline")
-  rescue TWEnv::Error::NoSuchArchiveError
+  rescue TWEnv::Error::ArchiveNotFoundError
     line.error("Can't continue because an archive for #{user} doesn't exist").end
   rescue Twitter::Error => ex
     line.end.error("#{ex.message} (#{ex.class})").end